Fix: SSR bundling

+++ runtimes/node/versions/latest/helpers/next-js/bundle.sh
@@ -4,12 +4,15 @@ if [ -n "$OPEN_RUNTIMES_OUTPUT_DIRECTORY" ]; then
     cd $OPEN_RUNTIMES_OUTPUT_DIRECTORY
 fi
 
-mkdir -p .next
-mv ./* .next/
+ENTRYPOINT="./server/webpack-runtime.js"
+if [ -e "$ENTRYPOINT" ]; then
+    mkdir -p .next
+    mv ./* .next/
 
-if [ -d "/usr/local/build/public/" ]; then
-    mv /usr/local/build/public/ ./public/
+    if [ -d "/usr/local/build/public/" ]; then
+        mv /usr/local/build/public/ ./public/
+    fi
+
+    mv /usr/local/build/package*.json ./
+    mv /usr/local/build/node_modules/ ./node_modules/
 fi
-
-mv /usr/local/build/package*.json ./
-mv /usr/local/build/node_modules/ ./node_modules/
