+++ runtimes/node/versions/latest/helpers/sveltekit/bundle.sh
@@ -4,5 +4,10 @@ if [ -n "$OPEN_RUNTIMES_OUTPUT_DIRECTORY" ]; then
     cd $OPEN_RUNTIMES_OUTPUT_DIRECTORY
 fi
 
-mv /usr/local/build/package*.json ./
-mv /usr/local/build/node_modules/ ./node_modules/
+ENTRYPOINT="./handler.js"
+if [ -e "$ENTRYPOINT" ]; then
+    echo "Bundling with server-side rendering support ..."
+    mv /usr/local/build/package*.json ./
+    mv /usr/local/build/node_modules/ ./node_modules/
+fi
+
